# schema/openapi.yaml
openapi: 3.0.0
servers:
  - url: 'http://localhost:5000/tpa/v1'
    description: Local development server
info:
  title: "fable+ Team Performance Scan Backend"
  description: |
    <h1>Default Responses</h1>

    Every route can, in addition to the defined responses for the specific
    route, return one of the default responses defined here.

    Responses that return with an HTTP status code other than 200 will
    additionally contain the `X-ErrorCode` header set to the
    corresponsing error code.

    <table>
        <thead>
            <tr>
                <th>Response Code</th>
                <th>Error</th>
                <th>Description</th>
            </tr>
        </thead>
        <tbody>
            <tr>
                <td>400</td>
                <td>1002</td>
                <td>JSON body is missing</td>
            </tr>
            <tr>
                <td>400</td>
                <td>1003</td>
                <td>JSON body is malformed</td>
            </tr>
            <tr>
                <td>400</td>
                <td>1005</td>
                <td>Bad parameter</td>
            </tr>
            <tr>
                <td>400</td>
                <td>1006</td>
                <td>Invalid header</td>
            </tr>
            <tr>
                <td>401</td>
                <td>1101</td>
                <td>Authorization header is missing</td>
            </tr>
            <tr>
...
            </tr>
...
security:
  - apiAuth: []
...
paths:
...
  /user/login:
    post:
      operationId: loginUser
      description: >-
        Route is used to authenticate a user with the API service and to obtain
        a session token. Authentication can either be by mail and password for
        registered users, or with an invitation ID.
      tags:
      - User
      security: []
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userLoginResponse'
        '403':
          description: Invalid credentials
          x-ErrorCode: 1203
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      requestBody:
        $ref: '#/components/requestBodies/userLoginRequest'
  /user/register:
    post:
      operationId: registerUser
      description: Route creates a new user.
      tags:
      - User
      responses:
        '200':
          description: User successfuly created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createdResponse'
        '400':
          description: Invalid password format
          x-ErrorCode: 1201
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        '409':
          description: E-Mail already registered
          x-ErrorCode: 1202
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      requestBody:
        $ref: '#/components/requestBodies/userRegistrationRequest'
  /user/update:
    post:
      operationId: updateUser
      description: Update personal data for user.
      tags:
      - User
      responses:
        '200':
          description: User data successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/successResponse'
        '400':
          description: Invalid password format
          x-ErrorCode: 1201
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      requestBody:
        $ref: '#/components/requestBodies/userUpdateRequest'
...
components:
  securitySchemes:
    apiAuth:
      type: http
      scheme: bearer
      description: |
        Authentication with API session token
      bearerFormat: API session token
...
  requestBodies:
    userLoginRequest:
      content:
        application/json:
          schema:
            $ref: user_login_req.json
    userRegistrationRequest:
      content:
        application/json:
          schema:
            $ref: user_registration_req.json
    userUpdateRequest:
      content:
        application/json:
          schema:
            $ref: user_update_req.json
...
  schemas:
    baseResponse:
        type: object
        properties: {}
    successResponse:
      allOf:
        - type: object
          properties:
            success:
              type: boolean
        - $ref: '#/components/schemas/baseResponse'
    createdResponse:
      allOf:
        - type: object
          properties:
            id:
              type: string
        - $ref: '#/components/schemas/baseResponse'
    errorResponse:
      allOf:
        - type: object
          properties:
            error:
              type: object
              properties:
                errorCode:
                  type: integer
                errorMessage:
                  type: string
        - $ref: '#/components/schemas/baseResponse'
...
    userLoginResponse:
      allOf:
        - $ref: '#/components/schemas/baseResponse'
        - $ref: 'user_login_resp.json'
...